backend/requirements.txt
Flask==2.3.2
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
tensorflow==2.13.0
